♻️ refactor: 웹훅 처리에 전략 패턴 도입

+++ src/main/java/com/gogym/gympay/entity/GymPay.java
@@ -23,7 +23,7 @@ public class GymPay extends BaseEntity {
   @OneToOne(mappedBy = "gymPay")
   private Member member;
 
-  public void charge(long amount) {
+  public void charge(int amount) {
     this.balance += amount;
   }
 }
